meeting-agent/requirements.txt
livekit-agents>=1.4.0
livekit-plugins-silero>=1.4.0
livekit-plugins-elevenlabs>=1.4.0
livekit-plugins-anthropic>=1.4.0
livekit-plugins-anam~=1.4.0
livekit>=1.0.0
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
supabase>=2.0.0
